test: add util to query by component type
